dev_requirement.txt
pytest>=3.0
pyinstaller>=3.2
pynsist>=1.12
django-modeltranslation>=0.12.1
